refactor - 서치페이지 리팩토링
